refactors get_links and variations
